\name{arrivals}
\alias{arrivals}
\docType{data}
\title{
International Arrivals to Australia
}
\description{
Quarterly international arrivals (in thousands) to Australia from Japan, New Zealand, UK and the US.  1981Q1 - 2012Q3.
}
\usage{data(arrivals)}
\format{Quarterly time series of class \code{ts}.}

\source{
Tourism Research Australia.
}
\examples{
autoplot(arrivals)
}
\keyword{datasets}
