\name{ausair}
\alias{ausair}
\docType{data}
\title{
Air Transport Passengers Australia
}
\description{
Total annual air passengers including domestic and international aircraft passengers of air carriers registered in Australia. 1970-2009.
}
\usage{data(ausair)}
\format{Annual time series of class \code{ts}.}

\source{
World Bank. \url{data.is/x5KiEO}
}
\examples{
plot(ausair)
}
\keyword{datasets}
