% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{ausbeer}
\alias{ausbeer}
\title{Quarterly Australian Beer production}
\format{
Quarterly time series of class \code{ts}.
}
\source{
Australian Bureau of Statistics. Cat. 8301.0.55.001.
}
\description{
Total quarterly beer production in Australia (in megalitres) from 1956:Q1 to
2010:Q2.
}
\examples{

data(ausbeer)
ggseasonplot(ausbeer)

}
\keyword{datasets}
