% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{auscafe}
\alias{auscafe}
\title{Monthly expenditure on eating out in Australia}
\format{Monthly time series of class \code{ts}.}
\source{
Australian Bureau of Statistics. Catalogue No. 8501.0
}
\description{
The total monthly expenditure on cafes, restaurants and takeaway food
services in Australia ($billion). April 1982 - September 2017.
}
\examples{

autoplot(auscafe)

}
\keyword{datasets}
