% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{austa}
\alias{austa}
\title{International visitors to Australia}
\format{
Annual time series of class \code{ts}.
}
\source{
International Visitor Survey, Tourism Research Australia.
}
\description{
Total international visitors to Australia (in millions). 1980-2015.
}
\examples{

autoplot(austa)

}
\keyword{datasets}
