% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{austourists}
\alias{austourists}
\title{International Tourists to Australia: Total visitor nights.}
\format{
Quarterly time series of class \code{ts}.
}
\source{
Tourism Research Australia.
}
\description{
Quarterly visitor nights (in millions) spent by international tourists to
Australia. 1999-2015.
}
\examples{

autoplot(austourists)

}
\keyword{datasets}
