\name{cafe}
\alias{cafe}
\docType{data}
\title{
Quarterly expenditure on eating out in Australia
}
\description{
The total quarterly expenditure on cafes, restaurants and takeway food services in Australia
(1982:Q2-2010:Q4)
}
\usage{data(cafe)}

\format{Annual time series of class \code{ts}.}

\source{Australian Bureau of Statistics. Catalogue No. 8501.0}

\examples{
plot(cafe)
}
\keyword{datasets}

