% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{calls}
\alias{calls}
\title{Call volume for a large North American bank}
\format{Time series object of class \code{msts}.}
\source{
Rob Hyndman
}
\description{
Five-minute call volume handled on weekdays between 7:00am and 9:05pm, from
3 March 2003 to 23 May 2003.
}
\examples{

autoplot(calls, xlab="Weeks")

}
\keyword{datasets}
