% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{elecdemand}
\alias{elecdemand}
\alias{elecdaily}
\title{Half-hourly and daily electricity demand for Victoria, Australia, in 2014}
\format{
Multiple time series of class \code{mts}.
}
\source{
Australian Energy Market Operator, and the Australian Bureau of
Meteorology.
}
\description{
\code{elecdemand} is a half-hourly time series matrix with three columns:
  \tabular{ll}{
    Demand:      \tab Total electricity demand in GW for Victoria, Australia, every half-hour during 2014.\cr
    WorkDay:     \tab taking value 1 on work days, and 0 otherwise.\cr
    Temperature: \tab half-hourly temperatures for Melbourne (BOM site 086071).
}
\code{elecdaily} is a daily time series matrix with three columns:
  \tabular{ll}{
    Demand:      \tab Total electricity demand in GW for Victoria, Australia, every day during 2014.\cr
    WorkDay:     \tab taking value 1 on work days, and 0 otherwise.\cr
    Temperature: \tab maximum daily temperatures for Melbourne (BOM site 086071).
}
}
\details{
This data is for operational demand, which is the demand met by local
scheduled generating units, semi-scheduled generating units, and
non-scheduled intermittent generating units of aggregate capacity larger
than 30 MW, and by generation imports to the region. The operational demand
excludes the demand met by non-scheduled non-intermittent generating units,
non-scheduled intermittent generating units of aggregate capacity smaller
than 30 MW, exempt generation (e.g. rooftop solar, gas tri-generation, very
small wind farms, etc), and demand of local scheduled loads. It also
excludes some very large industrial users (such as mines or smelters).
}
\examples{

summary(elecdemand)
summary(elecdaily)

}
\keyword{datasets}
