\name{elecequip}
\alias{elecequip}
\docType{data}
\title{
Electrical equipment manufactured in the Euro area.
}
\description{
Manufacture of electrical equipment: computer, electronic and optical products. Data adjusted by working days; Euro area (16 countries). Industry new orders index. 2005=100.
}
\usage{data(elecequip)}
\format{Time series object of class \code{ts}.
}
\source{
Eurostat. \url{data.is/y6dO8i}
}

\examples{
plot(elecequip)
}
\keyword{datasets}
