\name{elecsales}
\alias{elecsales}
\docType{data}
\title{
Electricity sales to residential customers in South Australia.
}
\description{
Annual electricity sales for South Australia in GWh from 1989 to 2008. Electricity used for hot water has been excluded.
}
\usage{data(elecsales)}
\format{Time series object of class \code{ts}.
}
\source{
Australian Energy Market Operator.
}

\examples{
plot(elecsales)
}
\keyword{datasets}
