% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euretail}
\alias{euretail}
\title{Quarterly retail trade: Euro area.}
\format{
Quarterly time series of class \code{ts}.
}
\source{
Eurostat. \url{http://data.is/IdKyZr}
}
\description{
Quarterly retail trade index in the Euro area (17 countries), 1996-2011,
covering wholesale and retail trade, and repair of motor vehicles and
motorcycles. (Index: 2005 = 100).
}
\examples{

autoplot(euretail)

}
\keyword{datasets}
