% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{package}
\name{fpp-package}
\alias{fpp-package}
\alias{fpp}
\alias{fpp2}
\title{Data for "Forecasting: principles and practice" (2nd ed, 2017)}
\description{
All data sets required for the examples and exercises in the book
"Forecasting: principles and practice" (2nd ed, 2017) by Rob J Hyndman and George
Athanasopoulos. This package also loads other packages required for the
book.
}
\seealso{
\code{\link[forecast]{forecast}} package
}
\author{
Rob J Hyndman

Maintainer: Rob.Hyndman@monash.edu
}
\keyword{package}
