% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp2.R
\docType{package}
\name{fpp2-package}
\alias{fpp2}
\alias{fpp2-package}
\title{fpp2: Data for "Forecasting: Principles and Practice" (2nd Edition)}
\description{
All data sets required for the examples and exercises
  in the book "Forecasting: principles and practice" (2nd ed, 2018)
  by Rob J Hyndman and George Athanasopoulos <https://otexts.com/fpp2/>.
  All packages required to run the examples are also loaded.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/fpp2-package/}
  \item \url{https://github.com/robjhyndman/fpp2-package}
  \item \url{https://otexts.com/fpp2/}
  \item Report bugs at \url{https://github.com/robjhyndman/fpp2-package/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} [copyright holder]

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
