% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fpp2_packages}
\alias{fpp2_packages}
\title{List all packages loaded by fpp2}
\usage{
fpp2_packages(include_self = FALSE)
}
\arguments{
\item{include_self}{Include fpp2 in the list?}
}
\value{
A character vector of package names.
}
\description{
List all packages loaded by fpp2
}
\examples{
fpp2_packages()
}
