\name{gasoline}
\alias{gasoline}
\docType{data}
\title{
US finished motor gasoline product supplied.
}
\description{Weekly data beginning 2 February 1991, ending 20 January 2017. Units are "thousand barrels per day".

}
\usage{data(gasoline)}
\format{Time series object of class \code{ts}.
}
\source{
US Energy Information Adminstration.
}

\examples{
plot(gasoline, xlab="Year")
}
\keyword{datasets}
