\name{goog}
\alias{goog}
\docType{data}
\title{
Daily closing stock prices of Google Inc
}
\description{Closing stock prices of GOOG from the NASDAQ exchange, for 1000 consecutive trading days between 25 February 2013 and 13 February 2017.
}
\usage{data(goog)}

\format{Daily time series of class \code{ts}.}

\source{\url{https://www.quandl.com/data/GOOG/NASDAQ_GOOGL}}

\examples{
autoplot(goog)
}
\keyword{datasets}

