% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{guinearice}
\alias{guinearice}
\title{Rice production (Guinea)}
\format{Annual time series of class \code{ts}.}
\source{
World Bank. \url{http://data.is/whKD8H}
}
\description{
Total annual rice production (million metric tons) for Guinea. 1970-2011.
}
\examples{

autoplot(guinearice)

}
\keyword{datasets}
