% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{h02}
\alias{h02}
\title{Monthly cortecosteroid drug sales in Australia from July 1991 to June 2008.}
\format{Monthly time series of class \code{ts}.}
\source{
Medicare Australia
}
\description{
Total monthly scripts for pharmaceutical products falling under ATC code
H02, as recorded by the Australian Health Insurance Commission. Measured in
millions of scripts.
}
\examples{

autoplot(h02)
ggseasonplot(h02)

}
\keyword{datasets}
