% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{livestock}
\alias{livestock}
\title{Livestock (sheep) in Asia, 1961-2007.}
\format{
Annual time series of class \code{ts}.
}
\source{
United Nations.
}
\description{
Annual sheep livestock numbers in Asia (in million head).
}
\examples{

autoplot(livestock)

}
\keyword{datasets}
