% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{marathon}
\alias{marathon}
\title{Boston marathon winning times since 1897}
\format{Annual time series of class \code{ts}.}
\source{
Boston Athletic Association.
\url{http://www.baa.org/races/boston-marathon/boston-marathon-history/champions/mens-open-division.aspx}
}
\description{
Winning times (in minutes) for the Boston Marathon Men's Open Division.
1897-2016.
}
\examples{

autoplot(marathon)

}
\keyword{datasets}
