% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{maxtemp}
\alias{maxtemp}
\title{Maximum annual temperatures at Moorabbin Airport, Melbourne}
\format{Annual time series of class \code{ts}.}
\source{
Australian Bureau of Meteorology.
}
\description{
Maximum annual temperatures (degrees Celsius) for Moorabbin Airport,
Melbourne. 1971-2016.
}
\examples{

autoplot(maxtemp)

}
\keyword{datasets}
