% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{mens400}
\alias{mens400}
\title{Winning times in Olympic men's 400m track final. 1896-2016.}
\format{
time series of class \code{ts} with frequency 1/4.
}
\description{
Times in seconds for the gold-medal winner of the men's 400m track final
at each Olympics since 1896. Missing values occur in 1916, 1940 and 1944
due to the World Wars.
}
\examples{

autoplot(mens400)

}
\keyword{datasets}
