\name{oil}
\alias{oil}
\docType{data}
\title{
Annual oil production in Saudi Arabia
}
\description{
Annual oil production (millions of tonnes), Saudi Arabia, 1965-2013.
}
\usage{data(oil)}

\format{Annual time series of class \code{ts}.}

\source{BP. \url{http://data.is/GEOHU3}}

\examples{
plot(oil)
}
\keyword{datasets}

