% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{qauselec}
\alias{qauselec}
\title{Quarterly Australian Electricity production}
\format{
Quarterly time series of class \code{ts}.
}
\source{
Australian Bureau of Statistics. Cat. 8301.0.55.001.
}
\description{
Total quarterly electricity production in Australia (in billion kWh) from
1956:Q1 to 2010:Q2.
}
\examples{

autoplot(qauselec)

}
\keyword{datasets}
