\name{qcement}
\alias{qcement}
\docType{data}
\title{
Quarterly Australian Portland Cement production
}
\description{
Total quarterly production of Portland cement in Australia (in millions of tonnes) from 1956:Q1 to 2014:Q1.
}
\usage{data(ausbeer)}
\format{Quarterly time series of class \code{ts}.}

\source{
Australian Bureau of Statistics. Cat. 8301.0.55.001.
}
\examples{
autoplot(qcement)
}
\keyword{datasets}
