\name{qauselec}
\alias{qauselec}
\docType{data}
\title{
Quarterly Australian Electricity production
}
\description{
Total quarterly electricity production in Australia (in billion kWh) from 1956:Q1 to 2010:Q2.
}
\usage{data(qauselec)}
\format{Quarterly time series of class \code{ts}.}

\source{
Australian Bureau of Statistics. Cat. 8301.0.55.001.
}
\examples{
plot(qauselec)
}
\keyword{datasets}
