% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qgas}
\alias{qgas}
\title{Quarterly Australian Gas Production}
\format{
Quarterly time series of class \code{ts}.
}
\source{
Australian Bureau of Statistics. Cat. 8301.0.55.001.
}
\description{
Total quarterly gas production in Australia (in petajoules) from 1956:Q1 to
2010:Q2.
}
\examples{

autoplot(qgas)

}
\keyword{datasets}
