\name{usconsumption}
\alias{usconsumption}
\docType{data}
\title{
Growth rates of personal consumption and personal income in the USA.
}
\description{
Percentage changes in quarterly personal consumption expenditure and personal disposable income for the US, 1970 to 2015.
}
\usage{data(usconsumption)}
\format{Time series object of class \code{ts}.
}
\source{
Federal Reserve Bank of St Louis. \url{http://data.is/AnVtzB}. \url{http://data.is/wQPcjU}.
}

\examples{
plot(usconsumption)
}
\keyword{datasets}
