% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usmelec}
\alias{usmelec}
\title{Electricity monthly total net generation. January 1973 - June 2013.}
\format{
Time series object of class \code{ts}.
}
\source{
US Energy Information Administration.
}
\description{
Electricity net generation measured in billions of kilowatt hours (kWh).
}
\examples{

autoplot(usmelec)

}
\keyword{datasets}
