\name{usmelec}
\alias{usmelec}
\docType{data}
\title{
Electricity monthly total net generation. January 1973 - June 2013.
}
\description{
Electricity net generation measured in billions of kilowatt hours (kWh).
}
\usage{data(usmelec)}
\format{Time series object of class \code{ts}.
}
\source{
US Energy Information Administration. \url{http://data.is/zgRWCO}.
}

\examples{
autoplot(usmelec)
}
\keyword{datasets}
