% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{visnights}
\alias{visnights}
\title{Quarterly visitor nights for various regions of Australia.}
\format{Time series object of class \code{mts}.}
\source{
Tourism Research Australia.
}
\description{
Total quarterly visitor nights (in millions) from 1998-2016 for twenty
regions of Australia within six states. The states are:
New South Wales, Queensland, South Australia, Victoria, Western Australia, and
Other.
}
\examples{

autoplot(visnights)

}
\keyword{datasets}
