% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{vn}
\alias{vn}
\title{Quarterly visitor nights for various regions of Australia.}
\format{Time series object of class \code{mts}.}
\source{
Tourism Research Australia.
}
\description{
Total quarterly visitor nights (in millions) from 1998-2015 for eight
regions of Australia:
  \tabular{ll}{
    Sydney     \tab The Sydney metropolitan area.\cr
    NSW        \tab New South Wales other than Sydney\cr
    Melbourne  \tab The Melbourne metropolitan area.\cr
    VIC        \tab Victoria other than Melbourne.\cr
    BrisbaneGC \tab The Brisbane and Gold Coast area.\cr
    QLD        \tab Queensland other than Brisbane and the Gold Coast.\cr
    Capitals   \tab The other five capital cities: Adelaide, Hobart, Perth, Darwin and Canberra.\cr
    Other      \tab All other areas of Australia.
}
}
\examples{

autoplot(vn, facets=TRUE)

}
\keyword{datasets}
