\name{a10}
\alias{a10}
\docType{data}
\title{
Monthly anti-diabetic drug sales in Australia from 1991 to 2008.
}
\description{
Total monthly scripts for pharmaceutical products falling under ATC code A10, as recorded by the Australian Health Insurance Commission. July 1991 - June 2008.
}
\usage{data(a10)}

\format{Monthly time series of class \code{ts}.}

\source{Medicare Australia}

\examples{
plot(a10)
seasonplot(a10)
}
\keyword{datasets}
