% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{arrivals}
\alias{arrivals}
\title{International Arrivals to Australia}
\format{Quarterly time series of class \code{ts}.}
\source{
Tourism Research Australia.
}
\description{
Quarterly international arrivals (in thousands) to Australia from Japan, New
Zealand, UK and the US.  1981Q1 - 2012Q3.
}
\examples{

autoplot(arrivals)

}
\keyword{datasets}
