% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ausair}
\alias{ausair}
\title{Air Transport Passengers Australia}
\format{
Annual time series of class \code{ts}.
}
\source{
World Bank. \url{http://data.is/x5KiEO}
}
\description{
Total annual air passengers (in millions) including domestic and
international aircraft passengers of air carriers registered in Australia.
1970-2016.
}
\examples{

autoplot(ausair)

}
\keyword{datasets}
