\name{ausair}
\alias{ausair}
\docType{data}
\title{
Air Transport Passengers Australia
}
\description{
Total annual air passengers (in millions) including domestic and international aircraft passengers of air carriers registered in Australia. 1970-2015.
}
\usage{data(ausair)}
\format{Annual time series of class \code{ts}.}

\source{
World Bank. \url{http://data.is/x5KiEO}
}
\examples{
autoplot(ausair)
}
\keyword{datasets}
