\name{austa}
\alias{austa}
\docType{data}
\title{
International vistors to Australia
}
\description{
Total international visitors to Australia (in millions). 1980-2015.
}
\usage{data(austa)}
\format{Annual time series of class \code{ts}.}

\source{
International Visitor Survey, Tourism Research Australia.
}
\examples{
plot(austa)
}
\keyword{datasets}
