\name{austourists}
\alias{austourists}
\docType{data}
\title{
International Tourists to Australia: Total visitor nights.
}
\description{
Quarterly visitor nights (in millions) spent by international tourists to Australia. 1999-2015.
}
\usage{data(austourists)}
\format{Quarterly time series of class \code{ts}.}

\source{
Tourism Research Australia.
}
\examples{
autoplot(austourists)
}
\keyword{datasets}
