\name{austourists}
\alias{austourists}
\docType{data}
\title{
International Tourists to Australia: Total visitor nights.
}
\description{
Quarterly visitor nights spent by international tourists to Australia. 1999-2010.
}
\usage{data(austourists)}
\format{Annual time series of class \code{ts}.}

\source{
Tourism Research Australia.
}
\examples{
plot(austourists)
}
\keyword{datasets}
