\name{auscafe}
\alias{auscafe}
\docType{data}
\title{
Monthly expenditure on eating out in Australia
}
\description{
The total monthly expenditure on cafes, restaurants and takeway food services in Australia ($billion). April 1982 - October 2016.
}
\usage{data(auscafe)}

\format{Monthly time series of class \code{ts}.}

\source{Australian Bureau of Statistics. Catalogue No. 8501.0}

\examples{
plot(auscafe)
}
\keyword{datasets}

