% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{debitcards}
\alias{debitcards}
\title{Retail debit card usage in Iceland.}
\format{
Monthly time series of class \code{ts}.
}
\source{
Statistics Iceland. \url{http://data.is/HFzL69}
}
\description{
Monthly retail debit card usage in Iceland (million ISK). January 2000 -
August 2013.
}
\examples{

autoplot(debitcards)

}
\keyword{datasets}
