\name{debitcards}
\alias{debitcards}
\docType{data}
\title{
Retail debit card usage in Iceland.
}
\description{
Retail debit card usage in Iceland (million ISK). January 2000 - August 2013.
}
\usage{data(debitcards)}

\format{Annual time series of class \code{ts}.}

\source{Statistics Iceland. \url{http://data.is/HFzL69}}

\examples{
plot(debitcards)
}
\keyword{datasets}

