% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departures}
\alias{departures}
\title{Total monthly departures from Australia}
\format{
Multiple monthly time series of class \code{mts} containing the
following series:
\tabular{ll}{
  permanent \tab permanent departures from Australia.\cr
  reslong   \tab long-term resident departures from Australia.\cr
  vislong   \tab long-term visitor departures from Australia.\cr
  resshort  \tab short-term resident departures from Australia.\cr
  visshort  \tab short-term visitor departures from Australia.
}
}
\source{
Australian Bureau of Statistics. Catalogue No 3401.02.
}
\description{
Overseas departures from Australia: permanent departures, long-term (more
than one year) residents departing, long-term (more than one year) visitors
departing, short-term (less than one year) residents departing and
short-term (less than one year) visitors departing. January 1976 - November
2016.
}
\examples{

autoplot(departures, facets=TRUE)

}
\keyword{datasets}
