\name{elecequip}
\alias{elecequip}
\docType{data}
\title{
Electrical equipment manufactured in the Euro area.
}
\description{
Monthly manufacture of electrical equipment: computer, electronic and optical products. January 1996 - March 2012. Data adjusted by working days; Euro area (17 countries). Industry new orders index. 2005=100.
}
\usage{data(elecequip)}
\format{Time series object of class \code{ts}.
}
\source{
Eurostat. \url{http://data.is/y6dO8i}
}

\examples{
plot(elecequip)
}
\keyword{datasets}
