% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elecequip}
\alias{elecequip}
\title{Electrical equipment manufactured in the Euro area.}
\format{
Time series object of class \code{ts}.
}
\source{
Eurostat. \url{http://data.is/y6dO8i}
}
\description{
Monthly manufacture of electrical equipment: computer, electronic and
optical products. January 1996 - March 2012. Data adjusted by working days;
Euro area (17 countries). Industry new orders index. 2005=100.
}
\examples{

autoplot(elecequip)

}
\keyword{datasets}
