% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elecsales}
\alias{elecsales}
\title{Electricity sales to residential customers in South Australia.}
\format{
Time series object of class \code{ts}.
}
\source{
Australian Energy Market Operator.
}
\description{
Annual electricity sales for South Australia in GWh from 1989 to 2008.
Electricity used for hot water has been excluded.
}
\examples{

autoplot(elecsales)

}
\keyword{datasets}
