\name{euretail}
\alias{euretail}
\docType{data}
\title{
Quarterly retail trade: Euro area.
}
\description{
Quarterly retail trade index in the Euro area (17 countries), 1996-2011, covering wholesale and retail trade, and repair of motor vehicles and motorcycles. (Index: 2005 = 100).
}
\usage{data(euretail)}

\format{Quarterly time series of class \code{ts}.}

\source{Eurostat. \url{http://data.is/IdKyZr}}

\examples{
plot(euretail)
}
\keyword{datasets}

