\name{fpp-package}
\alias{fpp-package}
\alias{fpp}
\docType{package}
\title{Data for "Forecasting: principles and practice"}
\description{All data sets required for the examples and exercises in the book "Forecasting: principles and practice" by Rob J Hyndman and George Athanasopoulos.
}
\details{
\tabular{ll}{
Package: \tab fpp\cr
Type: \tab Package\cr
License: \tab GPL2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Rob J Hyndman

Maintainer: Rob.Hyndman@monash.edu
}
\keyword{ package }
\seealso{
\code{\link[forecast]{forecast}} package
}
