% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{goog}
\alias{goog}
\title{Daily closing stock prices of Google Inc}
\format{Daily time series of class \code{ts}.}
\source{
\url{https://www.quandl.com/data/GOOG/NASDAQ_GOOGL}
}
\description{
Closing stock prices of GOOG from the NASDAQ exchange, for 1000 consecutive
trading days between 25 February 2013 and 13 February 2017.
}
\examples{

autoplot(goog)

}
\keyword{datasets}
