\name{guinearice}
\alias{guinearice}
\docType{data}
\title{
Rice production (Guinea)
}
\description{
Total annual rice production (million metric tons) for Guinea. 1970-2011.
}
\usage{data(guinearice)}
\format{Annual time series of class \code{ts}.}

\source{
World Bank. \url{http://data.is/whKD8H}
}
\examples{
plot(guinearice)
}
\keyword{datasets}
