% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{hyndsight}
\alias{hyndsight}
\title{Daily pageviews for the Hyndsight blog. 30 April 2014 to 29 April 2015.}
\format{Time series object of class \code{ts}.}
\source{
Rob Hyndman
}
\description{
Hyndsight is Rob Hyndman's personal blog at
\url{http://robjhyndman.com/hyndsight}. This series contains the daily
pageviews for one year, beginning 30 April 2014. The frequency is set to 7,
to allow the weekly pattern to be modelled.
}
\examples{

autoplot(hyndsight, xlab="Weeks")

}
\keyword{datasets}
