\name{insurance}
\alias{insurance}
\docType{data}
\title{
Insurance quotations and advertising expenditure.
}
\description{
Monthly quotations and monthly television advertising expenditure for a US insurance company.
January 2002 to April 2005.
}
\usage{data(insurance)}
\format{Monthly time series of class \code{ts}.}

\source{
Kindly provided by Dave Reilly, Automatic Forecasting Systems.
}
\examples{
autoplot(insurance)
}
\keyword{datasets}
