\name{livestock}
\alias{livestock}
\docType{data}
\title{
Livestock (sheep) in Asia, 1961-2007.
}
\description{
Annual sheep livestock  numbers in Asia (in million head).
}
\usage{data(livestock)}

\format{Annual time series of class \code{ts}.}

\source{United Nations. \url{http://data.is/GFxwQi}}

\examples{
plot(livestock)
}
\keyword{datasets}

