\name{marathon}
\alias{marathon}
\docType{data}
\title{Boston marathon winning times since 1897}

\description{
Winning times (in minutes) for the Boston Marathon Men's Open Division. 1897-2016.}
\usage{data(marathon)}
\format{Annual time series of class \code{ts}.}

\source{
Boston Athletic Association. \url{http://www.baa.org/races/boston-marathon/boston-marathon-history/champions/mens-open-division.aspx}
}
\examples{
autoplot(marathon)
}
\keyword{datasets}
