\name{maxtempe}
\alias{maxtemp}
\docType{data}
\title{
Maximum annual temperatures at Moorabbin Airport, Melbourne
}
\description{
Maximum annual temperatures (degrees Celsius) for Moorabbin Airport, Melbourne. 1971-2016.
}
\usage{data(maxtemp)}
\format{Annual time series of class \code{ts}.}

\source{
Australian Bureau of Meterology.
}
\examples{
plot(maxtemp)
}
\keyword{datasets}
