% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{oil}
\alias{oil}
\title{Annual oil production in Saudi Arabia}
\format{
Annual time series of class \code{ts}.
}
\source{
BP. \url{http://data.is/GEOHU3}
}
\description{
Annual oil production (millions of tonnes), Saudi Arabia, 1965-2013.
}
\examples{

autoplot(oil)

}
\keyword{datasets}
