\name{qgas}
\alias{qgas}
\docType{data}
\title{
Quarterly Australian Gas Production
}
\description{
Total quarterly gas production in Australia (in petajoules) from 1956:Q1 to 2010:Q2.
}
\usage{data(qgas)}
\format{Quarterly time series of class \code{ts}.}

\source{
Australian Bureau of Statistics. Cat. 8301.0.55.001.
}
\examples{
plot(qgas)
}
\keyword{datasets}
