% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{sunspotarea}
\alias{sunspotarea}
\title{Annual average sunspot area (1875-2015)}
\format{
Annual time series of class \code{ts}.
}
\source{
NASA
}
\description{
Annual averages of the daily sunspot areas (in units of millionths of a
hemisphere) for the full sun. Sunspots are magnetic regions that appear as
dark spots on the surface of the sun. The Royal Greenwich Observatory
compiled daily sunspot observations from May 1874 to 1976. Later data are
from the US Air Force and the US National Oceanic and Atmospheric
Administration. The data have been calibrated to be consistent across the
whole history of observations.
}
\examples{

autoplot(sunspotarea)

}
\keyword{datasets}
