% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp-package.R
\docType{data}
\name{wmurders}
\alias{wmurders}
\title{Annual female murder rate (per 100,000 standard population) in the USA.
1950-2004.}
\format{Annual time series of class \code{ts}.}
\source{
Gapminder Foundation. \url{http://data.is/XKa24F}
}
\description{
Total Murdered women, per 100 000 standard population.
}
\examples{

autoplot(wmurders)

}
\keyword{datasets}
