\name{wmurders}
\alias{wmurders}
\docType{data}
\title{
Annual female murder rate (per 100,000 standard population) in the USA.
}
\description{
Total Murdered women, per 100 000 standard population.
}
\usage{data(wmurders)}

\format{Annual time series of class \code{ts}.}

\source{Gapminder Foundation. \url{http://data.is/XKa24F}}

\examples{
plot(wmurders)
}
\keyword{datasets}
